requirements.txt
streamlit==1.31.0
google-generativeai==0.3.2
python-dotenv==1.0.0
PyPDF2==3.0.1
Pillow>=10.0.0
streamlit-webrtc==0.47.1
av==10.0.0
pydub==0.25.1
SpeechRecognition==3.10.0
requests==2.31.0
